+++ packages/react-wildcat-handoff/src/utils/simpleRender.js
@@ -2,10 +2,11 @@
 
 const defaultTemplate = require("./defaultTemplate.js");
 
-module.exports = function serverRender(cfg) {
+module.exports = function serverRender({
+    request,
+    wildcatConfig
+}) {
     let result = {};
-
-    const wildcatConfig = cfg.wildcatConfig;
     const htmlTemplate = wildcatConfig.serverSettings.htmlTemplate || defaultTemplate;
 
     const html = htmlTemplate({
@@ -25,6 +26,7 @@ module.exports = function serverRender(cfg) {
             }
         },
         html: "",
+        request,
         wildcatConfig
     });