+++ packages/react-wildcat/src/utils/startWebSocketServer.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const WebSocketServer = require("ws").Server;
+const WebSocketServer = require("uws").Server;
 const chokidar = require("chokidar");
 
 function send(eventName, response, client) {
@@ -10,11 +10,11 @@ function send(eventName, response, client) {
     }));
 }
 
-module.exports = function startWebSocketServer(root, options) {
-    const cache = options.cache;
-    const server = options.server;
-    const watchOptions = options.watchOptions;
-
+module.exports = function startWebSocketServer(root, {
+    cache,
+    server,
+    watchOptions
+}) {
     const wss = new WebSocketServer({
         server
     });